# app/docs/swagger.yaml
consumes:
- application/json
definitions:
  apperrors.ErrorCode:
    enum:
    - access_token_expired
    - all_signals_failed_processing
    - authentication_error
    - authorization_error
    - client_closed
    - database_error
    - forbidden
    - invalid_correlation_id
    - internal_error
    - invalid_request
    - invalid_url_param
    - malformed_body
    - not_implemented
    - network_error
    - password_too_short
    - refresh_token_invalid
    - request_too_large
    - unsupported_media_type
    - rate_limit_exceeded
    - resource_already_exists
    - resource_expired
    - resource_in_use
    - resource_not_found
    - timeout
    - token_creation_failed
    type: string
    x-enum-varnames:
    - ErrCodeAccessTokenExpired
    - ErrCodeAllSignalsFailedProcessing
    - ErrCodeAuthenticationFailure
    - ErrCodeAuthorizationFailure
    - ErrCodeClientClosed
    - ErrCodeDatabaseError
    - ErrCodeForbidden
    - ErrCodeInvalidCorrelationID
    - ErrCodeInternalError
    - ErrCodeInvalidRequest
    - ErrCodeInvalidURLParam
    - ErrCodeMalformedBody
    - ErrCodeNotImplemented
    - ErrCodeNetworkError
    - ErrCodePasswordTooShort
    - ErrCodeRefreshTokenInvalid
    - ErrCodeRequestTooLarge
    - ErrCodeUnsupportedMediaType
    - ErrCodeRateLimitExceeded
    - ErrCodeResourceAlreadyExists
    - ErrCodeResourceExpired
    - ErrCodeResourceInUse
    - ErrCodeResourceNotFound
    - ErrCodeTimeout
    - ErrCodeFailedToCreateToken
...
  handlers.CreateIsnRequest:
    properties:
      detail:
        example: Sample ISN description
        type: string
      is_in_use:
        example: true
        type: boolean
      title:
        example: Sample ISN @example.org
        type: string
      visibility:
        enum:
        - public
        - private
        example: private
        type: string
    type: object
  handlers.CreateIsnResponse:
    properties:
      id:
        example: 67890684-3b14-42cf-b785-df28ce570400
        type: string
      resource_url:
        example: http://localhost:8080/api/isn/sample-isn
        type: string
      slug:
        example: sample-isn
        type: string
    type: object
...
  handlers.Isn:
    properties:
      created_at:
        example: "2025-06-03T13:47:47.331787+01:00"
        type: string
      detail:
        example: Sample ISN description
        type: string
      id:
        example: 67890684-3b14-42cf-b785-df28ce570400
        type: string
      is_in_use:
        example: true
        type: boolean
      slug:
        example: sample-isn
        type: string
      title:
        example: Sample ISN @example.org
        type: string
      updated_at:
        example: "2025-06-03T13:47:47.331787+01:00"
        type: string
      user_account_id:
        example: a38c99ed-c75c-4a4a-a901-c9485cf93cf3
        type: string
      visibility:
        enum:
        - public
        - private
        example: private
        type: string
    type: object
...
  handlers.IsnAndLinkedInfo:
    properties:
      isn:
        $ref: '#/definitions/handlers.Isn'
      signal_types:
        items:
          $ref: '#/definitions/handlers.SignalType'
        type: array
      user:
        $ref: '#/definitions/handlers.User'
    type: object
...
  handlers.SignalType:
    properties:
      content_kind:
        enum:
        - json
        - document
        - event
        example: json
        type: string
      created_at:
        example: "2025-06-03T13:47:47.331787+01:00"
        type: string
      detail:
        example: Sample signal type description
        type: string
      id:
        example: 67890684-3b14-42cf-b785-df28ce570400
        type: string
      is_in_use:
        example: true
        type: boolean
      readme_url:
        example: https://example.com/readme.md
        type: string
      schema_url:
        example: https://example.com/schema.json
        type: string
      sem_ver:
        example: ""
        type: string
      slug:
        example: sample-signal-type
        type: string
      title:
        example: Sample Signal Type
        type: string
      updated_at:
        example: "2025-06-03T13:47:47.331787+01:00"
        type: string
    type: object
...
  handlers.TransferIsnOwnershipRequest:
    properties:
      new_owner_account_id:
        example: a38c99ed-c75c-4a4a-a901-c9485cf93cf3
        type: string
    type: object
...
  handlers.UpdateIsnRequest:
    properties:
      detail:
        example: Sample ISN description
        type: string
      is_in_use:
        example: true
        type: boolean
      visibility:
        enum:
        - public
        - private
        example: private
        type: string
    type: object
...
  handlers.User:
    properties:
      account_id:
        example: a38c99ed-c75c-4a4a-a901-c9485cf93cf3
        type: string
      created_at:
        example: "2025-06-03T13:47:47.331787+01:00"
        type: string
      updated_at:
        example: "2025-06-03T13:47:47.331787+01:00"
        type: string
    type: object
...
  responses.ErrorResponse:
    properties:
      error_code:
        allOf:
        - $ref: '#/definitions/apperrors.ErrorCode'
        description: Standard app error code
        example: error_code
      message:
        description: Client message
        example: client message
        type: string
      request_id:
        description: request tracking id
        example: a1b2c3d4
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Signals ISN service API for managing Information Sharing Networks

    ## Common Error Responses
    All endpoints may return:
    - `400` Malformed request (invalid json, missing required fields, etc.)
    - `401` Unauthorized (invalid credentials)
    - `403` Forbidden (insufficient permissions)
    - `404` Resource not found (e.g. an ISN or signal type that doesn't exist or is not in use)
    - `413` Request body exceeds size limit
    - `429` Rate limit exceeded
    - `500` Internal server error
    - `504` Request timed out (`timeout`)

    Individual endpoints document their specific business logic errors.

    Standard error responses include a JSON response body containing `error_code`, `message` and `request_id` fields.

    The /oauth endpoints return `error` (the RFC6749 standard error code), `error_code` and `error_description` fields instead.

    ## Request Limits
    The endpoints are protected by:
    - **Rate limiting**: Configurable requests per second (all endpoints)
    - **Request size limits** (configurable - these are the defaults): 64KB for admin/auth endpoints, 5MB for json and event signal submissions, 20MB for document uploads

    Check the `Signalsd-Max-Request-Size` response header for the configured size limit of the endpoint called.

    The rate limit is set globally and prevents abuse of the service.
...

    ## Authentication & Authorization

    ### OAuth
...

...

...
    ```
...
    ```

...
paths:
...
  /api/admin/isn/{isn_slug}/transfer-ownership:
    put:
      description: |-
        Transfer ownership of an ISN to another admin account.
        This can be used when an admin leaves or when reorganizing responsibilities.
        Only site admins can transfer ISN ownership.
      parameters:
      - description: ISN slug
        example: sample-isn
        in: path
        name: isn_slug
        required: true
        type: string
      - description: Transfer details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.TransferIsnOwnershipRequest'
      responses:
        "200":
          description: OK
        "400":
          description: malformed_body | invalid_request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: authorization_error | access_token_expired
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: resource_not_found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: database_error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Transfer ISN Ownership
      tags:
      - ISN Configuration
...
  /api/isn:
    get:
      description: get a list of the configured ISNs (any authenticated account can
        use this endpoint - private ISNs are included)
      parameters:
      - default: false
        description: Include inactive ISNs
        in: query
        name: include_inactive
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.Isn'
            type: array
        "401":
          description: authorization_error | access_token_expired
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: database_error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Get ISN configurations
      tags:
      - ISN Configuration
  /api/isn/:
    post:
      description: |-
        Create an Information Sharing Network (ISN)

        *Visibility*
        Signals in a private ISN can only be seen by members of the ISN.
        Signals in a public ISN can be viewed by anyone (no authentication required).
        Accounts need write permission for the ISN before they can submit signals

        Once the ISN is created you can:
        - Grant accounts permission to use it
        - Add the Signal Types that can be shared over the network

        There can be multiple ISNs on the site and each ISN can have a
        different membership and be configured for different Signal Types.

        This endpoint can only be used by ISN admins and site admins

        All fields (title, detail, is_in_use and visibility) are required.
      parameters:
      - description: ISN configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateIsnRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateIsnResponse'
        "400":
          description: malformed_body
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: authorization_error | access_token_expired
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: resource_already_exists
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: database_error | internal_error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Create an ISN
      tags:
      - ISN Configuration
  /api/isn/{isn_slug}:
    get:
      description: Returns details about the ISN (any authenticated account can use
        this endpoint)
      parameters:
      - description: ISN slug
        example: sample-isn
        in: path
        name: isn_slug
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.IsnAndLinkedInfo'
        "401":
          description: authorization_error | access_token_expired
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: resource_not_found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: database_error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Get an ISN configuration
      tags:
      - ISN Configuration
    put:
      description: |-
        Update the ISN configuration
        This endpoint can only be used by admin accounts
        ISN admins can only update ISNs they own
        Site admins can update any ISN
      parameters:
      - description: ISN slug
        example: sample-isn
        in: path
        name: isn_slug
        required: true
        type: string
      - description: ISN configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateIsnRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: malformed_body
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: authorization_error | access_token_expired
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: resource_not_found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: database_error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Update an ISN
      tags:
      - ISN Configuration
...
produces:
...
securityDefinitions:
  BearerAccessToken:
    description: Bearer {JWT access token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...